# CUDA: fix overflow in FA, tune performance