Due to numerical overflows the CUDA FlashAttention code on master does not work correctly for very long contexts (something like several million tokens across all sequences). This PR uses 64 bit math for those parts of the code susceptible to such problems: the K/V offsets between sequences and the calculation of K/V offsets within a sequence. For the vector kernel there was a performance regression on Pascal when simply casting the offsets to 64 bit, for this reason I'm adding a 32 bit offset after each iteration (turns out to be faster for Pascal/AMD anyways). I am not seeing any performance differences for the other kernels so I'm just casting the offsets to 64 bit. While working on this I noticed that at some point the tile FA kernels seem to have gotten faster than the vector kernels on my RX 6800 so I'm enabling them for AMD.

<details>
<summary>Performance changes</summary>

| GPU       | Model           |     Microbatch size | Test     |     t/s master |     t/s ec05b081e |     Speedup |
| :-------- | :-------------- | ------------------: | :------- | -------------: | ----------------: | ----------: |
| P40       | llama 8B Q4_0   |                   1 | pp8192   |          45.19 |             46.21 |        1.02 |
| P40       | llama 8B Q4_0   |                   2 | pp8192   |          79.03 |             81.57 |        1.03 |
| P40       | llama 8B Q4_0   |                   4 | pp8192   |         100.38 |            103.55 |        1.03 |
| P40       | llama 8B Q4_0   |                   8 | pp8192   |         111.95 |            115.22 |        1.03 |
| P40       | llama 8B Q4_0   |                  16 | pp8192   |         326.43 |            327.61 |        1.00 |
| P40       | llama 8B Q4_0   |                  32 | pp8192   |         458.59 |            459.75 |        1.00 |
| P40       | llama 8B Q4_0   |                  64 | pp8192   |         519.87 |            521.12 |        1.00 |
| P40       | llama 8B Q4_0   |                 128 | pp8192   |         560.72 |            564.46 |        1.01 |
| P40       | llama 8B Q4_0   |                 256 | pp8192   |         595.92 |            598.70 |        1.00 |
| P40       | llama 8B Q4_0   |                 512 | pp8192   |         608.42 |            610.60 |        1.00 |
| P40       | llama 8B Q4_0   |                1024 | pp8192   |         599.67 |            604.16 |        1.01 |
| P40       | llama 8B Q4_0   |                2048 | pp8192   |         581.61 |            580.96 |        1.00 |
| P40       | llama 8B Q4_0   |                4096 | pp8192   |         578.04 |            583.86 |        1.01 |
| P40       | llama 8B Q4_0   |                8192 | pp8192   |         578.77 |            582.35 |        1.01 |
| RTX 3090  | llama 8B Q4_0   |                   1 | pp8192   |         140.13 |            139.89 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                   2 | pp8192   |         249.16 |            250.09 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                   4 | pp8192   |         425.18 |            426.44 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                   8 | pp8192   |         529.60 |            528.76 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                  16 | pp8192   |        1132.49 |           1131.78 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                  32 | pp8192   |        1842.52 |           1841.17 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                  64 | pp8192   |        2807.42 |           2798.67 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                 128 | pp8192   |        3534.20 |           3516.30 |        0.99 |
| RTX 3090  | llama 8B Q4_0   |                 256 | pp8192   |        4192.59 |           4172.36 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                 512 | pp8192   |        4426.59 |           4397.17 |        0.99 |
| RTX 3090  | llama 8B Q4_0   |                1024 | pp8192   |        4529.86 |           4496.75 |        0.99 |
| RTX 3090  | llama 8B Q4_0   |                2048 | pp8192   |        4494.54 |           4483.05 |        1.00 |
| RTX 3090  | llama 8B Q4_0   |                4096 | pp8192   |        4509.37 |           4485.62 |        0.99 |
| RTX 3090  | llama 8B Q4_0   |                8192 | pp8192   |        4490.27 |           4481.76 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                   1 | pp8192   |         168.07 |            167.63 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                   2 | pp8192   |         302.88 |            303.42 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                   4 | pp8192   |         591.02 |            592.69 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                   8 | pp8192   |        1005.46 |           1006.44 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                  16 | pp8192   |        1692.36 |           1690.25 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                  32 | pp8192   |        3107.26 |           3105.39 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                  64 | pp8192   |        5421.35 |           5433.31 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                 128 | pp8192   |        7964.31 |           7995.78 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                 256 | pp8192   |       10339.18 |          10333.54 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                 512 | pp8192   |       11580.34 |          11574.64 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                1024 | pp8192   |       11811.83 |          11801.82 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                2048 | pp8192   |       11432.54 |          11399.22 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                4096 | pp8192   |       11411.55 |          11403.61 |        1.00 |
| RTX 4090  | llama 8B Q4_0   |                8192 | pp8192   |       11417.33 |          11400.58 |        1.00 |
| RX 6800   | llama 8B Q4_0   |                   1 | pp8192   |          38.91 |             46.38 |        1.19 |
| RX 6800   | llama 8B Q4_0   |                   2 | pp8192   |          68.44 |             67.47 |        0.99 |
| RX 6800   | llama 8B Q4_0   |                   4 | pp8192   |          73.64 |             73.36 |        1.00 |
| RX 6800   | llama 8B Q4_0   |                   8 | pp8192   |          75.06 |             75.16 |        1.00 |
| RX 6800   | llama 8B Q4_0   |                  16 | pp8192   |          86.76 |             93.01 |        1.07 |
| RX 6800   | llama 8B Q4_0   |                  32 | pp8192   |          94.86 |            115.82 |        1.22 |
| RX 6800   | llama 8B Q4_0   |                  64 | pp8192   |         101.16 |            127.03 |        1.26 |
| RX 6800   | llama 8B Q4_0   |                 128 | pp8192   |         113.84 |            153.90 |        1.35 |
| RX 6800   | llama 8B Q4_0   |                 256 | pp8192   |         118.68 |            161.18 |        1.36 |
| RX 6800   | llama 8B Q4_0   |                 512 | pp8192   |         118.96 |            159.27 |        1.34 |
| RX 6800   | llama 8B Q4_0   |                1024 | pp8192   |         116.08 |            148.57 |        1.28 |
| RX 6800   | llama 8B Q4_0   |                2048 | pp8192   |         106.82 |            134.90 |        1.26 |
| RX 6800   | llama 8B Q4_0   |                4096 | pp8192   |         106.94 |            134.76 |        1.26 |
| RX 6800   | llama 8B Q4_0   |                8192 | pp8192   |         106.93 |            135.25 |        1.26 |


</details>