@he29-net 2026-01-06
Hi! I recently updated to a newer ROCm (7.0.3) and llama.cpp release and ran into an issue with increased VRAM usage. `git bisect` led me here (where, to my surprise, I found my own comments about an unrelated performance issue.. :D)

The new regression is that before this commit, the VRAM consumption with FA enabled is constant. With:
`./llama-server -m ../../../models/Meta-Llama-3.1-8B-Instruct-Q8_0.gguf -ngl 99 -c 0 -fa -ctk q8_0 -ctv q5_0`,
`htop` shows constant VRAM usage of 15413 MB while processing a long text file.

After rebuilding with this commit, the VRAM usage keeps increasing, and around 20k tokens in, llama.cpp runs out of memory. Tweaking parameters to decrease VRAM usage only delay the failure – at the rate of memory usage growth, it seems that using the full context window would need extra 6 to 8 GB of memory than previously.

Looking at the log, FA shows up as enabled, and the rate of the increase does not seem as steep as when FA is disabled either, so that does not seem to be the issue. So I just wanted to make sure if it is an expected consequence of this change, of whether it could be some sort of memory leak (perhaps "triggered" by the new ROCm version) that may need to be looked into?

Thanks! Martin